Separate out a hook
Used for title & title-button visibility in the home component.
